Move the SoftLinkChange test to an integration test.
Refactor to use testutil methods, simplifying the test.
