+++ internal/mtail/mtail_test.go
@@ -44,97 +44,6 @@ func startMtailServer(t *testing.T, options ...func(*Server) error) *Server {
 	return m
 }
 
-func TestHandleSoftLinkChange(t *testing.T) {
-	if testing.Short() {
-		t.Skip("skipping test in short mode")
-	}
-	workdir, rmWorkdir := testutil.TestTempDir(t)
-	defer rmWorkdir()
-
-	logFilepath := path.Join(workdir, "log")
-	m := startMtailServer(t, LogPathPatterns(logFilepath))
-	defer m.Close(true)
-
-	trueLog1, err := os.Create(logFilepath + ".true1")
-	if err != nil {
-		t.Errorf("could not create log file: %s", err)
-	}
-	defer trueLog1.Close()
-	err = os.Symlink(logFilepath+".true1", logFilepath)
-	if err != nil {
-		t.Errorf("could not create symlink: %s", err)
-	}
-	inputLines := []string{"hi1", "hi2", "hi3"}
-	for _, x := range inputLines {
-		testutil.WriteString(t, trueLog1, x+"\n")
-		testutil.FatalIfErr(t, trueLog1.Sync())
-	}
-	check3 := func() (bool, error) {
-		if expvar.Get("log_count").String() != "1" {
-			return false, nil
-		}
-		if expvar.Get("log_rotations_total").(*expvar.Map).Get(logFilepath) != nil {
-			return false, nil
-		}
-		return true, nil
-	}
-	ok, err := testutil.DoOrTimeout(check3, 1*time.Second, 10*time.Millisecond)
-	testutil.FatalIfErr(t, err)
-	if !ok {
-		t.Errorf("log count: received %s, expected 1", expvar.Get("log_count").String())
-		t.Errorf("log rotatins: received %s, expected 0", expvar.Get("log_rotations_total").String())
-	}
-	trueLog2, err := os.Create(logFilepath + ".true2")
-	if err != nil {
-		t.Errorf("could not create log file: %s", err)
-	}
-	defer trueLog2.Close()
-	err = os.Remove(logFilepath)
-	if err != nil {
-		t.Errorf("could not delete symlink: %s", err)
-	}
-	err = os.Symlink(logFilepath+".true2", logFilepath)
-	if err != nil {
-		t.Errorf("could not create symlink: %s", err)
-	}
-	for _, x := range inputLines {
-		testutil.WriteString(t, trueLog2, x+"\n")
-		testutil.FatalIfErr(t, trueLog2.Sync())
-	}
-	check6 := func() (bool, error) {
-		if expvar.Get("log_count").String() != "1" {
-			return false, nil
-		}
-		if expvar.Get("log_rotations_total").(*expvar.Map).Get(logFilepath) == nil {
-			return false, nil
-		}
-		if expvar.Get("log_rotations_total").(*expvar.Map).Get(logFilepath).String() != "1" {
-			return false, nil
-		}
-		return true, nil
-	}
-	ok, err = testutil.DoOrTimeout(check6, 100*time.Millisecond, 10*time.Millisecond)
-	if err != nil {
-		buf := make([]byte, 1<<16)
-		count := runtime.Stack(buf, true)
-		t.Log("Timed out: Dumping goroutine stack")
-		t.Log(string(buf[:count]))
-		t.Fatal(err)
-	}
-	if !ok {
-		t.Errorf("log count: received %s, expected 1", expvar.Get("log_count").String())
-		t.Errorf("log rotatins: received %s, expected 0", expvar.Get("log_rotations_total").String())
-	}
-	_, err = os.Stat(logFilepath + ".true1")
-	if err != nil {
-		t.Errorf("stat failed on %s: %s", logFilepath+".true1", err)
-	}
-	_, err = os.Stat(logFilepath + ".true2")
-	if err != nil {
-		t.Errorf("stat failed on %s: %s", logFilepath+".true2", err)
-	}
-}
-
 func TestGlob(t *testing.T) {
 	if testing.Short() {
 		t.Skip("skipping test in short mode")
